# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API DOCUMENTATION",
        "description": "This is the API documentation for the Video Streaming Chrome Extension.",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "https://chromex.adeleke.tech/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/": {},
        "/videos": {
            "get": {
                "tags": [
                    "Endpoints"
                ],
                "summary": "Retrieve all videos",
                "description": "Retrieve all videos.",
                "operationId": "getVideos",
                "responses": {
                    "200": {
                        "description": "Videos retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Videos retrieved successfully."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Video-1"
                                                    },
                                                    "public_url": {
                                                        "type": "string",
                                                        "example": "http://example.com/storage/videos/87uiytbHGf-1/video.webm"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No videos found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No videos found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/video/{id}": {
            "get": {
                "tags": [
                    "Endpoints"
                ],
                "summary": "Retrieve a video",
                "description": "Retrieve a video.",
                "operationId": "getVideo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Video ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Video retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Video retrieved successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Video-1"
                                                },
                                                "public_url": {
                                                    "type": "string",
                                                    "example": "http://example.com/storage/videos/87uiytbHGf-1/video.webm"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Video not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Video not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/video/stream": {
            "post": {
                "tags": [
                    "Endpoints"
                ],
                "summary": "Stream a video",
                "description": "Start streaming a video.",
                "operationId": "streamVideo",
                "requestBody": {
                    "description": "Video details",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "blob": {
                                        "type": "file",
                                        "format": "webm"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Video streamed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Video streamed successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Video already completed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Video already completed."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/video/end-stream/{title}": {
            "get": {
                "tags": [
                    "Endpoints"
                ],
                "summary": "Stop streaming a video",
                "description": "Stop streaming a video.",
                "operationId": "stopVideo",
                "parameters": [
                    {
                        "name": "title",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Video-1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Video completed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Video completed successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Video-1"
                                                },
                                                "url": {
                                                    "type": "string",
                                                    "example": "http://example.com/storage/videos/87uiytbHGf.webm"
                                                },
                                                "transcript": {
                                                    "type": "string",
                                                    "example": "This is the transcript of the video."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Video already completed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Video already completed."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Video not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Video not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}